batch: Move baseline replacement choices
The batch merge code has a dedicated baseline edit fallback and a separate
merge candidate enumerator. Both paths coordinate reviewed
replacement-origin placement when source anchors cannot decide a target
location.

The replacement-origin choice value object and enumeration logic still live
in the edit fallback module. That couples candidate discovery to edit
application and makes the shared placement decision look like an edit-only
concern.

This commit addresses that by moving replacement-origin choice data and
enumeration into a baseline replacement choices module, then importing that
helper from both the fallback editor and candidate enumerator.

Subsequent commits will update the architecture tests so the new ownership
boundary stays recorded.
