+++ src/git_stage_batch/batch/baseline_edits.py
@@ -3,8 +3,6 @@
 from __future__ import annotations
 
 from collections.abc import Iterable, Iterator, Sequence
-from dataclasses import dataclass
-import hashlib
 from typing import TYPE_CHECKING, Any
 
 from ..core.line_selection import LineRanges, LineSelection, coerce_line_ranges
@@ -14,6 +12,9 @@ from ..utils.text import normalize_line_endings
 from .baseline_reference_positions import (
     baseline_reference_absence_position as _find_baseline_absence_position,
     baseline_reference_insertion_position as _find_baseline_insertion_position,
+)
+from .baseline_replacement_choices import (
+    replacement_origin_choices_for_unit as _replacement_origin_choices_for_unit,
 )
 from .line_sequence_equality import (
     line_sequences_equal as _line_sequences_match,
@@ -30,41 +31,11 @@ _BaselineLineEdit = tuple[int, int, list[bytes]]
 _DEFAULT_RESOLUTION_CHOICE_LIMIT = 51
 
 
-@dataclass(frozen=True)
-class ReplacementOriginChoice:
-    """Concrete target placement for an origin-tracked replacement."""
-
-    choice_index: int
-    position: int
-    target_after_line: int | None
-    target_before_line: int | None
-
-
 def _selection_outside_bounds(lines: LineSelection, max_line: int) -> bool:
     for line in lines:
         if line < 1 or line > max_line:
             return True
     return False
-
-
-def _replacement_origin_ambiguity_key(
-    unit_index: int,
-    deletion_index: int,
-    origin: Any,
-    claimed_lines: Sequence[int],
-    forbidden_sequence: Sequence[bytes],
-) -> str:
-    claimed = ",".join(str(line) for line in claimed_lines)
-    digest = _sequence_digest(forbidden_sequence)
-    return (
-        f"replacement-origin:{unit_index}:delete:{deletion_index}:"
-        f"claimed:{claimed}:old:{origin.old_start}-{origin.old_end}:"
-        f"new:{origin.new_start}-{origin.new_end}:{digest}"
-    )
-
-
-def _sequence_digest(lines: Sequence[bytes]) -> str:
-    return hashlib.sha256(b"".join(lines)).hexdigest()[:12]
 
 
 def _baseline_removal_edit(
@@ -211,7 +182,7 @@ def _replacement_edit_from_origin_resolution(
     if resolution is None:
         return None
 
-    key, choices = replacement_origin_choices_for_unit(
+    key, choices = _replacement_origin_choices_for_unit(
         claim,
         unit_index,
         unit,
@@ -321,65 +292,6 @@ def _has_complete_baseline_references(
         if reference is None or not getattr(reference, "has_after_line", False):
             return False
     return bool(presence_line_set or deletion_claims)
-
-
-def replacement_origin_choices_for_unit(
-    claim: AbsenceClaim,
-    unit_index: int,
-    unit: Any,
-    claimed_lines: Sequence[int],
-    working_lines: Sequence[bytes],
-    *,
-    max_results: int | None = None,
-) -> tuple[str | None, tuple[ReplacementOriginChoice, ...]]:
-    """Return explicit target placements for an origin-tracked replacement."""
-    origin = getattr(unit, "origin", None)
-    if origin is None or not claim.content_lines:
-        return None, ()
-
-    forbidden_sequence = [
-        normalize_line_endings(line)
-        for line in claim.content_lines
-    ]
-    if not forbidden_sequence:
-        return None, ()
-    if len(forbidden_sequence) > len(working_lines):
-        return None, ()
-
-    choices: list[ReplacementOriginChoice] = []
-    for position in range(0, len(working_lines) - len(forbidden_sequence) + 1):
-        if not _line_slice_matches(working_lines, position, forbidden_sequence):
-            continue
-        choices.append(
-            ReplacementOriginChoice(
-                choice_index=len(choices) + 1,
-                position=position,
-                target_after_line=None if position == 0 else position,
-                target_before_line=(
-                    None
-                    if position + len(forbidden_sequence) >= len(working_lines)
-                    else position + len(forbidden_sequence) + 1
-                ),
-            )
-        )
-        if max_results is not None and len(choices) >= max_results:
-            break
-
-    if not choices:
-        return None, ()
-
-    deletion_indices = getattr(unit, "deletion_indices", [])
-    if len(deletion_indices) != 1:
-        return None, ()
-
-    key = _replacement_origin_ambiguity_key(
-        unit_index,
-        deletion_indices[0],
-        origin,
-        claimed_lines,
-        forbidden_sequence,
-    )
-    return key, tuple(choices)
 
 
 def try_apply_baseline_replacement_units(
